WeatherService 일부 메서드 GeoService로 리팩토링
